+++ dfdx-core/src/tensor_ops/concat_tensor_along/mod.rs
@@ -1,5 +1,11 @@
-pub use super::concat_tensor_along::ConcatAlongKernel;
+use super::concat_shape_along::TryConcatShapeAlong;
 use crate::{shapes::*, tensor::*};
+
+pub(crate) mod cpu_kernel;
+#[cfg(feature = "cuda")]
+pub(crate) mod cuda_kernel;
+#[cfg(feature = "webgpu")]
+mod webgpu_kernel;
 
 /// Concatenate two tensors along a given axis.
 ///
@@ -43,36 +49,54 @@ use crate::{shapes::*, tensor::*};
 /// let b: Tensor<(Const<2>, usize), f32, _> = dev.zeros_like(&(Const, 4));
 /// let _: Tensor<Rank2<2, 6>, f32, _> = (a, b).concat_along(Axis::<1>).realize();
 /// ```
-#[deprecated = "Use TryConcatTensorAlong or TryConcatShapeAlong instead"]
-pub trait TryConcatAlong<Ax>: Sized {
+pub trait TryConcatTensorAlong<Ax>: Sized {
     type Output;
 
     /// Concatenates self along the given axis.
-    fn concat_along(self, ax: Ax) -> Self::Output {
-        self.try_concat_along(ax).unwrap()
+    fn concat_tensor_along(self, ax: Ax) -> Self::Output {
+        self.try_concat_tensor_along(ax).unwrap()
     }
     /// Fallibly concatenates self along the given axis.
-    fn try_concat_along(self, ax: Ax) -> Result<Self::Output, Error>;
+    fn try_concat_tensor_along(self, ax: Ax) -> Result<Self::Output, Error>;
 }
 
-#[allow(deprecated)]
-impl<A, B, Ax, E: Dtype, D, T: Tape<E, D>, R: Tape<E, D>> TryConcatAlong<Ax>
+pub trait ConcatAlongKernel<E: Dtype>: Storage<E> {
+    fn forward<A: Shape, B: Shape, C: Shape>(
+        &self,
+        ax: usize,
+        a: &Tensor<A, E, Self>,
+        b: &Tensor<B, E, Self>,
+        c: &mut Tensor<C, E, Self>,
+    ) -> Result<(), Error>;
+
+    fn backward<A: Shape, B: Shape>(
+        &self,
+        ax: usize,
+        a: &GhostTensor<A, E, Self>,
+        grad_a: &mut Self::Vec,
+        b: &GhostTensor<B, E, Self>,
+        grad_b: &mut Self::Vec,
+        grad_out: &Self::Vec,
+    ) -> Result<(), Error>;
+}
+
+impl<A, B, Ax, E: Dtype, D, T: Tape<E, D>, R: Tape<E, D>> TryConcatTensorAlong<Ax>
     for (Tensor<A, E, D, T>, Tensor<B, E, D, R>)
 where
     Ax: Axes<Array = [isize; 1]>,
     D: ConcatAlongKernel<E> + ZerosTensor<E>,
     A: Shape + HasAxes<Ax>,
     B: Shape<Concrete = A::Concrete> + HasAxes<Ax>,
-    (A, B): TryConcatAlong<Ax>,
-    <(A, B) as TryConcatAlong<Ax>>::Output: Shape,
+    (A, B): TryConcatShapeAlong<Ax>,
+    <(A, B) as TryConcatShapeAlong<Ax>>::Output: Shape,
     T: Merge<R>,
 {
-    type Output = Tensor<<(A, B) as TryConcatAlong<Ax>>::Output, E, D, T>;
+    type Output = Tensor<<(A, B) as TryConcatShapeAlong<Ax>>::Output, E, D, T>;
 
-    fn try_concat_along(self, ax: Ax) -> Result<Self::Output, Error> {
+    fn try_concat_tensor_along(self, ax: Ax) -> Result<Self::Output, Error> {
         let (lhs, rhs) = self;
 
-        let out_shape = (*lhs.shape(), *rhs.shape()).concat_along(ax);
+        let out_shape = (*lhs.shape(), *rhs.shape()).concat_shape_along(ax);
         let ax = Ax::as_array()[0] as usize;
 
         let (lhs, tape) = lhs.split_tape();
@@ -98,70 +122,7 @@ where
     }
 }
 
-macro_rules! impl_concat {
-    ($Ax:expr, $NumDims:expr, [$($Head:tt),*], [$($Tail:tt),*]) => {
-        #[allow(deprecated)]
-        impl<A: Dim, B: Dim, $($Head: Dim, )* $($Tail: Dim, )*> TryConcatAlong<Axis<$Ax>>
-            for (
-                ($($Head, )* A, $($Tail, )*),
-                ($($Head, )* B, $($Tail, )*),
-            )
-        where
-            A: std::ops::Add<B>,
-            <A as std::ops::Add<B>>::Output: Dim,
-            {
-                type Output = (
-                    $($Head, )*
-                    <A as std::ops::Add<B>>::Output,
-                    $($Tail, )*
-                );
-
-                fn try_concat_along(self, _: Axis<$Ax>) -> Result<Self::Output, Error> {
-                    let (lhs, rhs) = self;
-                    let lhs_dims = lhs.concrete();
-                    let rhs_dims = rhs.concrete();
-                    for i in 0..$NumDims {
-                        if i != $Ax {
-                            assert_eq!(lhs_dims[i], rhs_dims[i]);
-                        }
-                    }
-                    let mut out_dims = lhs_dims;
-                    out_dims[$Ax] += rhs_dims[$Ax];
-                    Ok(Self::Output::from_concrete(&out_dims).unwrap())
-                }
-            }
-    };
-}
-
-impl_concat!(0, 1, [], []);
-impl_concat!(0, 2, [], [D1]);
-impl_concat!(0, 3, [], [D1, D2]);
-impl_concat!(0, 4, [], [D1, D2, D3]);
-impl_concat!(0, 5, [], [D1, D2, D3, D4]);
-impl_concat!(0, 6, [], [D1, D2, D3, D4, D5]);
-
-impl_concat!(1, 2, [D0], []);
-impl_concat!(1, 3, [D0], [D2]);
-impl_concat!(1, 4, [D0], [D2, D3]);
-impl_concat!(1, 5, [D0], [D2, D3, D4]);
-impl_concat!(1, 6, [D0], [D2, D3, D4, D5]);
-
-impl_concat!(2, 3, [D0, D1], []);
-impl_concat!(2, 4, [D0, D1], [D3]);
-impl_concat!(2, 5, [D0, D1], [D3, D4]);
-impl_concat!(2, 6, [D0, D1], [D3, D4, D5]);
-
-impl_concat!(3, 4, [D0, D1, D2], []);
-impl_concat!(3, 5, [D0, D1, D2], [D4]);
-impl_concat!(3, 6, [D0, D1, D2], [D4, D5]);
-
-impl_concat!(4, 5, [D0, D1, D2, D3], []);
-impl_concat!(4, 6, [D0, D1, D2, D3], [D5]);
-
-impl_concat!(5, 6, [D0, D1, D2, D3, D4], []);
-
 #[cfg(test)]
-#[allow(deprecated)]
 mod tests {
     use super::*;
     use crate::{tensor_ops::*, tests::*};
@@ -179,7 +140,7 @@ mod tests {
             .clone()
             .try_realize::<(usize, Const<3>, Const<4>)>()
             .unwrap();
-        let c = (a_dyn, b_dyn).concat_along(Axis::<0>);
+        let c = (a_dyn, b_dyn).concat_tensor_along(Axis::<0>);
         let c = c.try_realize::<(Const<5>, Const<3>, Const<4>)>().unwrap();
         let a_arr = a.array();
         let b_arr = b.array();
@@ -212,7 +173,7 @@ mod tests {
             .clone()
             .try_realize::<(Const<2>, usize, Const<4>)>()
             .unwrap();
-        let c = (a_dyn, b_dyn).concat_along(Axis::<1>);
+        let c = (a_dyn, b_dyn).concat_tensor_along(Axis::<1>);
         let c = c.try_realize::<(Const<2>, Const<5>, Const<4>)>().unwrap();
         let a_arr = a.array();
         let b_arr = b.array();
@@ -244,7 +205,7 @@ mod tests {
             .clone()
             .try_realize::<(Const<2>, Const<3>, usize)>()
             .unwrap();
-        let c = (a_dyn, b_dyn).concat_along(Axis::<2>);
+        let c = (a_dyn, b_dyn).concat_tensor_along(Axis::<2>);
         let c = c.try_realize::<(Const<2>, Const<3>, Const<5>)>().unwrap();
         let a_arr = a.array();
         let b_arr = b.array();
@@ -263,35 +224,5 @@ mod tests {
         let b_grads = b.leaky_trace().exp().sum().backward();
         assert_close_to_tensor!(concat_grads.get(&a), a_grads.get(&a));
         assert_close_to_tensor!(concat_grads.get(&b), b_grads.get(&b));
-    }
-
-    #[test]
-    fn test_concat_shape() {
-        let a: (usize, Const<5>) = (5, Const);
-        let b: (usize, Const<5>) = (3, Const);
-        assert_eq!((a, b).concat_along(Axis::<0>), (8, Const::<5>));
-
-        let a: (Const<5>, Const<5>) = (Const, Const);
-        let b: (usize, Const<5>) = (3, Const);
-        assert_eq!((a, b).concat_along(Axis::<0>), (8, Const::<5>));
-
-        let a: (usize, Const<5>) = (5, Const);
-        let b: (Const<3>, Const<5>) = (Const, Const);
-        assert_eq!((a, b).concat_along(Axis::<0>), (8, Const::<5>));
-
-        #[cfg(feature = "nightly")]
-        {
-            let a: (Const<5>, Const<5>) = (Const, Const);
-            let b: (Const<3>, Const<5>) = (Const, Const);
-            assert_eq!((a, b).concat_along(Axis::<0>), (Const::<8>, Const::<5>));
-        }
-    }
-
-    #[test]
-    #[should_panic = "left: 10\n right: 7"]
-    fn test_concat_shape_fails() {
-        let a = (5, 10);
-        let b = (3, 7);
-        (a, b).concat_along(Axis::<0>);
     }
 }
