+++ dfdx/tests/issue_tests.rs
@@ -18,12 +18,12 @@ fn test_issue_891() {
 
     impl<Input, const AXIS: isize> Module<Input> for ConcatTensorAlong<Axis<AXIS>>
     where
-        Input: TryConcatAlong<Axis<AXIS>>,
+        Input: TryConcatTensorAlong<Axis<AXIS>>,
     {
-        type Output = <Input as TryConcatAlong<Axis<AXIS>>>::Output;
+        type Output = <Input as TryConcatTensorAlong<Axis<AXIS>>>::Output;
 
         fn try_forward(&self, x: Input) -> Result<Self::Output, Error> {
-            x.try_concat_along(Axis)
+            x.try_concat_tensor_along(Axis)
         }
     }
 
@@ -32,13 +32,5 @@ fn test_issue_891() {
     let dev = Cpu::default();
     let x = dev.tensor([1.]);
     let m = dev.build_module::<f32>(Arch::default());
-    let y = m.forward(x);
-    /*
-    error[E0275]: overflow evaluating the requirement `((_, _, _, _), (..., ..., ..., ...)): dfdx::prelude::TryConcatAlong<...>`
-      --> dfdx/tests/issue_tests.rs:36:15
-       |
-    36 |     let y = m.forward(x);
-       |               ^^^^^^^
-       |
-        */
+    let _y: Tensor<Rank1<2>, _, _, _> = m.forward(x);
 }
